requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
astroid==3.0.0
async-timeout==4.0.3
attrs==23.1.0
causal-learn==0.1.3.6
cdt==0.6.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
contourpy==1.1.1
cryptography==41.0.4
cvxpy==1.3.2
cycler==0.12.0
Cython==0.29.36
dill==0.3.7
Django==4.2.5
django-cors-headers==4.2.0
django-decouple==2.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
dowhy==0.10.1
ecos==2.0.12
filelock==3.12.4
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.2
GPUtil==1.4.0
graphviz==0.20.1
idna==3.4
imageio==2.31.5
imbalanced-learn==0.11.0
importlib-resources==6.1.0
isort==5.12.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy-loader==0.3
lightgbm==4.1.0
lime==0.2.0.1
MarkupSafe==2.1.3
matplotlib==3.7.3
mccabe==0.7.0
mpld3==0.5.9
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.4
osqp==0.6.3
packaging==23.1
pandas==2.0.3
patsy==0.5.3
Pillow==10.0.1
plotly
platformdirs==3.11.0
# plotly==5.17.0
psycopg2-binary==2.9.7
pycparser==2.21
pydot==1.4.2
pylint==3.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyWavelets==1.4.1
qdldl==0.1.7.post0
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.10.1
scs==3.2.3
six==1.16.0
skrebate==0.62
sqlparse==0.4.4
statsmodels==0.14.0
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.7.10
tomli==2.0.1
tqdm==4.66.1
torch==2.1.0
tomlkit==0.12.1
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
yarl==1.9.2
zipp==3.17.0
clarabel==0.6.0
colorama==0.4.6
pybind11==2.11.1
